1083: Move edit QA comment view from cases to comments
