+++ accessibility_monitoring_platform/apps/cases/templates/cases/forms/qa_add_comment.html
@@ -3,7 +3,6 @@
 {% block title %}{{ case.organisation_name }} | Add comment{% endblock %}
 
 {% block content %}
-
 <div class="govuk-width-container">
     <div class="govuk-breadcrumbs">
         <ol class="govuk-breadcrumbs__list">
@@ -63,5 +62,4 @@
         </div>
     </main>
 </div>
-
 {% endblock %}
