+++ accessibility_monitoring_platform/apps/cases/urls.py
@@ -20,7 +20,6 @@ from accessibility_monitoring_platform.apps.cases.views import (
     CaseReportDetailsUpdateView,
     CaseQAProcessUpdateView,
     QACommentCreateView,
-    QACommentUpdateView,
     CaseReportCorrespondenceUpdateView,
     CaseReportFollowupDueDatesUpdateView,
     CaseNoPSBResponseUpdateView,
@@ -84,11 +83,6 @@ urlpatterns: List[URLPattern] = [
         "<int:case_id>/add-qa-comment/",
         login_required(QACommentCreateView.as_view()),
         name="add-qa-comment",
-    ),
-    path(
-        "<int:pk>/edit-qa-comment/",
-        login_required(QACommentUpdateView.as_view()),
-        name="edit-qa-comment",
     ),
     path(
         "<int:pk>/edit-contact-details/",
